Renamed module to "ReactionModel" and indented code
